Split surface.pip socket witness into TerminalController+ControlSurfacePip.swift
TerminalController+ControlSurfaceContext3.swift would cross the 500-line
tracked threshold in the speculative merge with main (main independently
grew it and split off Context4), so the PiP witness gets its own file,
matching the existing ContextN split pattern.
